Move LSP<->Rascal converters to appropriate package.
