+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/rascal/conversion/CallHierarchy.java
@@ -24,7 +24,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package org.rascalmpl.vscode.lsp.util;
+package org.rascalmpl.vscode.lsp.rascal.conversion;
 
 import com.google.gson.JsonPrimitive;
 import java.util.HashMap;
