+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/rascal/conversion/DocumentChanges.java
@@ -24,7 +24,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package org.rascalmpl.vscode.lsp.util;
+package org.rascalmpl.vscode.lsp.rascal.conversion;
 
 import java.util.ArrayList;
 import java.util.HashMap;
