+++ rascal-lsp/src/test/java/engineering/swat/rascal/lsp/util/DocumentChangesTest.java
@@ -43,7 +43,7 @@ import org.eclipse.lsp4j.WorkspaceEdit;
 import org.junit.Test;
 import org.rascalmpl.util.locations.ColumnMaps;
 import org.rascalmpl.values.IRascalValueFactory;
-import org.rascalmpl.vscode.lsp.util.DocumentChanges;
+import org.rascalmpl.vscode.lsp.rascal.conversion.DocumentChanges;
 import org.rascalmpl.vscode.lsp.util.locations.Locations;
 
 import io.usethesource.vallang.IConstructor;
